PROJECT/requirements.txt
asgiref==3.4.1
Django==3.2.6
django-admin-sortable2==1.0
django-appconf==1.0.4
django-json-widget==1.1.1
django-select2==7.7.1
future==0.18.2
pytz==2021.1
sqlparse==0.4.1
